Refactor: Split vote_instruction.rs into multiple files
